[Refactor]#4 QuerySearchModel을 data패키지로 이동
